requirements.txt
easyocr==1.7.2
pillow==12.0.0
redis==5.0.7
pytz==2024.1
aiohttp==3.10.5
aiogram==3.13.1
APScheduler==3.10.4
uvicorn==0.30.6
fastapi==0.109.1
pydantic
numpy
kurigram==2.2.15
tgcrypto2==1.3.4
nest-asyncio==1.6.0